+++ R/class_database.R
@@ -166,6 +166,7 @@ database_class <- R6::R6Class(
       file.rename(from = tmp, to = self$path)
     },
     produce_line = function(row) {
+      withr::local_options(.new = list(OutDec = "."))
       paste(map_chr(row, self$produce_subline), collapse = database_sep_outer)
     },
     produce_subline = function(element) {